output/promrw: Move one package up then drop the useless pkg package
