+++ internal/output/prometheusrw/remotewrite/trend.go
