+++ internal/output/prometheusrw/sigv4/const.go
