+++ internal/output/prometheusrw/sigv4/sigv4.go
