+++ internal/output/prometheusrw/sigv4/util_test.go
