+++ internal/output/prometheusrw/stale/stale.go
